+++ CHANGELOG.md
@@ -1,14 +1,14 @@
-## 0.3.0 - In Development
+## 0.3.0 - 2016-09-15
 
 * [#61](https://github.com/adamcaudill/yawast/issues/61) - SSL Session Count: force 3DES suites
 * [#23](https://github.com/adamcaudill/yawast/issues/23) - Add check for HTTP to HTTPS redirect
 * [#63](https://github.com/adamcaudill/yawast/issues/63) - Rename `--sslsessioncount` to `--tdessessioncount`
 
-## 0.2.2 - 2016-08-07
+## 0.2.2 - 2016-09-07
 
 * [#55](https://github.com/adamcaudill/yawast/issues/55) - Add Protocol Intolerance information. 
 * Update `ssllabs` required version to 1.24.0 to correct issue with new SSL Labs API release.
 
-## 0.2.1 - 2016-08-03
+## 0.2.1 - 2016-09-03
 
 * Initial Public Release